@@ -1,26 +1,26 @@
 {
-  "_from": "ansi-regex@^2.0.0",
-  "_id": "ansi-regex@2.1.1",
+  "_from": "ansi-regex@^3.0.0",
+  "_id": "ansi-regex@3.0.0",
   "_inBundle": false,
-  "_integrity": "sha1-w7M6te42DYbg5ijwRorn7yfWVN8=",
+  "_integrity": "sha1-7QMXwyIGT3lGbAKWa922Bas32Zg=",
   "_location": "/ansi-regex",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "ansi-regex@^2.0.0",
+    "raw": "ansi-regex@^3.0.0",
     "name": "ansi-regex",
     "escapedName": "ansi-regex",
-    "rawSpec": "^2.0.0",
+    "rawSpec": "^3.0.0",
     "saveSpec": null,
-    "fetchSpec": "^2.0.0"
+    "fetchSpec": "^3.0.0"
   },
   "_requiredBy": [
     "/strip-ansi"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-2.1.1.tgz",
-  "_shasum": "c3b33ab5ee360d86e0e628f0468ae7ef27d654df",
-  "_spec": "ansi-regex@^2.0.0",
+  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-3.0.0.tgz",
+  "_shasum": "ed0317c322064f79466c02966bddb605ab37d998",
+  "_spec": "ansi-regex@^3.0.0",
   "_where": "C:\\Users\\westb\\Desktop\\Diagram_Designer\\node_modules\\strip-ansi",
   "author": {
     "name": "Sindre Sorhus",
@@ -34,11 +34,11 @@
   "deprecated": false,
   "description": "Regular expression for matching ANSI escape codes",
   "devDependencies": {
-    "ava": "0.17.0",
-    "xo": "0.16.0"
+    "ava": "*",
+    "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
@@ -72,37 +72,14 @@
     "pattern"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "name": "Sindre Sorhus",
-      "email": "sindresorhus@gmail.com",
-      "url": "sindresorhus.com"
-    },
-    {
-      "name": "Joshua Appelman",
-      "email": "jappelman@xebia.com",
-      "url": "jbnicolai.com"
-    },
-    {
-      "name": "JD Ballard",
-      "email": "i.am.qix@gmail.com",
-      "url": "github.com/qix-"
-    }
-  ],
   "name": "ansi-regex",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/chalk/ansi-regex.git"
   },
   "scripts": {
-    "test": "xo && ava --verbose",
+    "test": "xo && ava",
     "view-supported": "node fixtures/view-codes.js"
   },
-  "version": "2.1.1",
-  "xo": {
-    "rules": {
-      "guard-for-in": 0,
-      "no-loop-func": 0
-    }
-  }
+  "version": "3.0.0"
 }
