@@ -1,26 +1,26 @@
 {
-  "_from": "is-fullwidth-code-point@^1.0.0",
-  "_id": "is-fullwidth-code-point@1.0.0",
+  "_from": "is-fullwidth-code-point@^2.0.0",
+  "_id": "is-fullwidth-code-point@2.0.0",
   "_inBundle": false,
-  "_integrity": "sha1-754xOG8DGn8NZDr4L95QxFfvAMs=",
+  "_integrity": "sha1-o7MKXE8ZkYMWeqq5O+764937ZU8=",
   "_location": "/is-fullwidth-code-point",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "is-fullwidth-code-point@^1.0.0",
+    "raw": "is-fullwidth-code-point@^2.0.0",
     "name": "is-fullwidth-code-point",
     "escapedName": "is-fullwidth-code-point",
-    "rawSpec": "^1.0.0",
+    "rawSpec": "^2.0.0",
     "saveSpec": null,
-    "fetchSpec": "^1.0.0"
+    "fetchSpec": "^2.0.0"
   },
   "_requiredBy": [
     "/string-width"
   ],
-  "_resolved": "https://registry.npmjs.org/is-fullwidth-code-point/-/is-fullwidth-code-point-1.0.0.tgz",
-  "_shasum": "ef9e31386f031a7f0d643af82fde50c457ef00cb",
-  "_spec": "is-fullwidth-code-point@^1.0.0",
+  "_resolved": "https://registry.npmjs.org/is-fullwidth-code-point/-/is-fullwidth-code-point-2.0.0.tgz",
+  "_shasum": "a3b30a5c4f199183167aaab93beefae3ddfb654f",
+  "_spec": "is-fullwidth-code-point@^2.0.0",
   "_where": "C:\\Users\\westb\\Desktop\\Diagram_Designer\\node_modules\\string-width",
   "author": {
     "name": "Sindre Sorhus",
@@ -31,17 +31,14 @@
     "url": "https://github.com/sindresorhus/is-fullwidth-code-point/issues"
   },
   "bundleDependencies": false,
-  "dependencies": {
-    "number-is-nan": "^1.0.0"
-  },
   "deprecated": false,
   "description": "Check if the character represented by a given Unicode code point is fullwidth",
   "devDependencies": {
-    "ava": "0.0.4",
-    "code-point-at": "^1.0.0"
+    "ava": "*",
+    "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
@@ -71,7 +68,10 @@
     "url": "git+https://github.com/sindresorhus/is-fullwidth-code-point.git"
   },
   "scripts": {
-    "test": "node test.js"
+    "test": "xo && ava"
   },
-  "version": "1.0.0"
+  "version": "2.0.0",
+  "xo": {
+    "esnext": true
+  }
 }
