@@ -1,28 +1,29 @@
 {
-  "_from": "string-width@^1.0.1",
-  "_id": "string-width@1.0.2",
+  "_from": "string-width@^2.0.0",
+  "_id": "string-width@2.1.1",
   "_inBundle": false,
-  "_integrity": "sha1-EYvfW4zcUaKn5w0hHgfisLmxB9M=",
+  "_integrity": "sha512-nOqH59deCq9SRHlxq1Aw85Jnt4w6KvLKqWVik6oA9ZklXLNIOlqg4F2yrT1MVaTjAqvVwdfeZ7w7aCvJD7ugkw==",
   "_location": "/string-width",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "string-width@^1.0.1",
+    "raw": "string-width@^2.0.0",
     "name": "string-width",
     "escapedName": "string-width",
-    "rawSpec": "^1.0.1",
+    "rawSpec": "^2.0.0",
     "saveSpec": null,
-    "fetchSpec": "^1.0.1"
+    "fetchSpec": "^2.0.0"
   },
   "_requiredBy": [
-    "/gauge",
-    "/wide-align"
+    "/ansi-align",
+    "/boxen",
+    "/widest-line"
   ],
-  "_resolved": "https://registry.npmjs.org/string-width/-/string-width-1.0.2.tgz",
-  "_shasum": "118bdf5b8cdc51a2a7e70d211e07e2b0b9b107d3",
-  "_spec": "string-width@^1.0.1",
-  "_where": "C:\\Users\\westb\\Desktop\\Diagram_Designer\\node_modules\\gauge",
+  "_resolved": "https://registry.npmjs.org/string-width/-/string-width-2.1.1.tgz",
+  "_shasum": "ab93f27a8dc13d28cac815c462143a6d9012ae9e",
+  "_spec": "string-width@^2.0.0",
+  "_where": "C:\\Users\\westb\\Desktop\\Diagram_Designer\\node_modules\\boxen",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -33,9 +34,8 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "code-point-at": "^1.0.0",
-    "is-fullwidth-code-point": "^1.0.0",
-    "strip-ansi": "^3.0.0"
+    "is-fullwidth-code-point": "^2.0.0",
+    "strip-ansi": "^4.0.0"
   },
   "deprecated": false,
   "description": "Get the visual width of a string - the number of columns required to display it",
@@ -44,7 +44,7 @@
     "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
@@ -85,5 +85,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "1.0.2"
+  "version": "2.1.1"
 }
