@@ -1,15 +1,8 @@
 import { fabric } from "fabric";
 
 export class Shape {
-    // shape constructor that stores a shape and 
-    // its children are initialized as an empty array
-    Shape = shape => {
-        this.Shape = shape;
-        this.Children = [];
-    }
-
     // shape construct that stores a shape and initializes the children
-    Shape = (shape, shapeChildren) => {
+    constructor(shape, shapeChildren) {
        this.Shape = shape;
        this.Children = shapeChildren;
     }
@@ -31,7 +24,7 @@ export class Shape {
      }
 
     // push a child onto the shape children children
-    pushChild = child => {
+    pushChild (child) {
         if(children == undefined || children == null)
         alert("Error. No children to add to shape \n. \
                Please make sure there are shape children\n");
@@ -42,4 +35,6 @@ export class Shape {
     // pushes multiple children to the shape children
     pushChildren = children => children.forEach(child => this._shapeChildren.push(child));
 }
+
+
 
