@@ -4,8 +4,9 @@ export var colorMap = {
     "Medium Maturity" : "#FFFF00",
     "High Maturity" : "#8ADF92",
     "Out of Scope" : "#C6C6C6",
-    "Workstream" : "# 15009B",
-    "Portfolio" : "# 15009B",
-    "L1 Business Process" : "# 15009B",
-    "L2 Business Process" : "# 15009B"
+    "Workflow Level": "#15009B",
+    "Workstream" : "#338219",
+    "Portfolio" : "#d96818",
+    "L1 Business Process" : "#15009B",
+    "L2 Business Process" : "#4599d9"
 };
