@@ -1,7 +1,1 @@
-var obj = require("C:/Users/YOHAN-DESKTOP/Desktop/project_10-master/project_10-master/Test_App/Json_Data/json1.json");
-
-var content = obj.Content[0];
-
-var children = content.Children;
-
-console.log(obj);
+export var obj = require("./Test_App/Json_Data/json1.json");
